refactor(GlobalLayout, ContentLayout): `index`로 이름 수정
